% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicDose-package.R
\docType{package}
\name{epicDose-package}
\alias{epicDose-package}
\title{Build daily dose from EPIC EHR system}
\description{
Need dose and, if applicable, weight and age data
}
\author{
\strong{Maintainer}: Cole Beck \email{cole.beck@vumc.org} (\href{https://orcid.org/0000-0002-6849-6255}{ORCID})

}
