% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicDose.R
\name{epicDose}
\alias{epicDose}
\title{Main function}
\usage{
epicDose(
  drug,
  drug.columns = NULL,
  age.data = NULL,
  age.columns = NULL,
  age.limits = NULL,
  wgt.data = NULL,
  wgt.columns = NULL,
  min.obs = 3,
  max.obs = 10,
  earliest.date = NA,
  drugname = NULL,
  comments = TRUE
)
}
\arguments{
\item{drug}{data.frame; dose data}

\item{drug.columns}{list; map variables to drug data}

\item{age.data}{data.frame; age or date of birth data}

\item{age.columns}{list; map variables to age.data}

\item{age.limits}{numeric with length 1 or 2; age for patient is recorded for earliest observation.
This age is compared to age.limits and patient is removed if outside limits.}

\item{wgt.data}{data.frame; weight data}

\item{wgt.columns}{list; map variables to wgt.data}

\item{min.obs}{numeric; any patient with less observations than this is removed; defaults to 3}

\item{max.obs}{numeric; any patient with more observations than this is removed; defaults to 10}

\item{earliest.date}{date; any patient with data before this is removed}

\item{drugname}{character string; potentially used in partial string matching when combo drugs are present}

\item{comments}{T/F; include comments?}
}
\value{
data set
}
\description{
Build daily dose
}
